sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-identity-sdk"
version = "0.1.0"
description = "Python SDK for Agent Identity Layer MVP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyjwt[crypto]>=2.9.0"
]
